+++ changelog.md
@@ -41,3 +41,5 @@ The first release of Hammerstone
  - Code cleanup
  - Fixed issue where 'harvestables' were not generating correctly.
  - Added initial support for mobs
+ - Fixed issue where craftables were no longer showing in the UI panel, likely re-introducing an issue with craftables using resource groups failing
+ - Added support for 'defaultModelShouldOverrideResourceObject' in hs_buildable